Implemented (completely untested) slave uxpanel system.
Master uxpanel can handle login and such, and then when they access a service that is hosted on a slave uxpanel instance, it will add an authentication token and redirect them. Then user is authenticated on the new system and then can access the service. If user presses back to panel or anything, it takes user back to the master instance.