Critical fix to password system.
Old password system required admin password to be stored in plaintext and user passwords were not salted. Current system will use the PBKDF2 algorithm, by default with SHA-512 hashing algorithm, for greatly improved security; this is used for both admin password and user passwords.

Note: all passwords in the old format are still accepted, so changing the codebase should not be a problem. It is recommended, however, that you update your database so that the new passwords won't be too long. To seamlessly upgrade your database (with no data loss), use "ALTER TABLE uxpanel.accounts MODIFY COLUMN password VARCHAR(512);".

Passwords can still be hash beforehand and then used to create an account (this is useful, say, if you want user to submit password at order form but don't want to send password over email in plaintext, which is indeed a bad idea). With this method, see mkpasswd.php with the PBKDF2 format selected. Basically, use include/pbkdf2.php and the pbkdf2_create_hash function; if $result is the function's output, then store the string "*pbkdf2*$result" into the database.